+++ src/main/java/com/group13/EmployeeManager/repository/PayrollRepository.java
@@ -5,6 +5,6 @@ import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
 
 @Repository
-public interface PayrollRepository extends JpaRepository<Payroll, Integer> {
+public interface PayrollRepository extends JpaRepository<Payroll, Long> {
 
 }
